+++ application/src/main/java/org/opentripplanner/graph_builder/module/OsmBoardingLocationsModule.java
@@ -17,7 +17,6 @@ import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.framework.i18n.LocalizedString;
 import org.opentripplanner.graph_builder.model.GraphBuilderModule;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.graph.index.StreetIndex;
 import org.opentripplanner.routing.linking.VertexLinker;
 import org.opentripplanner.service.osminfo.OsmInfoGraphBuildService;
 import org.opentripplanner.service.osminfo.model.Platform;
@@ -39,7 +38,6 @@ import org.opentripplanner.street.search.TraverseMode;
 import org.opentripplanner.street.search.TraverseModeSet;
 import org.opentripplanner.transit.model.site.RegularStop;
 import org.opentripplanner.transit.model.site.StationElement;
-import org.opentripplanner.transit.service.TimetableRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -68,7 +66,6 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
   private final Graph graph;
 
   private final OsmInfoGraphBuildService osmInfoGraphBuildService;
-  private final TimetableRepository timetableRepository;
   private final VertexFactory vertexFactory;
 
   private VertexLinker linker;
@@ -76,12 +73,10 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
   @Inject
   public OsmBoardingLocationsModule(
     Graph graph,
-    OsmInfoGraphBuildService osmInfoGraphBuildService,
-    TimetableRepository timetableRepository
+    OsmInfoGraphBuildService osmInfoGraphBuildService
   ) {
     this.graph = graph;
     this.osmInfoGraphBuildService = osmInfoGraphBuildService;
-    this.timetableRepository = timetableRepository;
     this.vertexFactory = new VertexFactory(graph);
   }
 
@@ -89,8 +84,7 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
   public void buildGraph() {
     LOG.info("Improving boarding locations by checking OSM entities...");
 
-    StreetIndex streetIndex = graph.getStreetIndexSafe(timetableRepository.getSiteRepository());
-    this.linker = streetIndex.getVertexLinker();
+    this.linker = graph.getLinkerSafe();
     int successes = 0;
 
     for (TransitStopVertex ts : graph.getVerticesOfType(TransitStopVertex.class)) {
@@ -106,7 +100,7 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
       if (alreadyLinked) continue;
       // only connect transit stops that are not part of a pathway network
       if (!ts.hasPathways()) {
-        if (!connectVertexToStop(ts, streetIndex)) {
+        if (!connectVertexToStop(ts, graph)) {
           LOG.debug("Could not connect {} at {}", ts.getStop().getCode(), ts.getCoordinate());
         } else {
           successes++;
@@ -116,7 +110,7 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
     LOG.info("Found {} OSM references which match a stop's id or code", successes);
   }
 
-  private boolean connectVertexToStop(TransitStopVertex ts, StreetIndex index) {
+  private boolean connectVertexToStop(TransitStopVertex ts, Graph index) {
     if (connectVertexToNode(ts, index)) return true;
 
     if (connectVertexToWay(ts, index)) return true;
@@ -139,10 +133,10 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
    *
    * @return if the vertex has been connected
    */
-  private boolean connectVertexToArea(TransitStopVertex ts, StreetIndex index) {
+  private boolean connectVertexToArea(TransitStopVertex ts, Graph graph) {
     RegularStop stop = ts.getStop();
-    var nearbyAreaGroups = index
-      .getEdgesForEnvelope(getEnvelope(ts))
+    var nearbyAreaGroups = graph
+      .findEdges(getEnvelope(ts))
       .stream()
       .filter(AreaEdge.class::isInstance)
       .map(AreaEdge.class::cast)
@@ -180,11 +174,11 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
    *
    * @return if the vertex has been connected
    */
-  private boolean connectVertexToWay(TransitStopVertex ts, StreetIndex index) {
+  private boolean connectVertexToWay(TransitStopVertex ts, Graph graph) {
     var stop = ts.getStop();
     var nearbyEdges = new HashMap<Platform, List<Edge>>();
 
-    for (var edge : index.getEdgesForEnvelope(getEnvelope(ts))) {
+    for (var edge : graph.findEdges(getEnvelope(ts))) {
       osmInfoGraphBuildService
         .findPlatform(edge)
         .ifPresent(platform -> {
@@ -229,9 +223,9 @@ public class OsmBoardingLocationsModule implements GraphBuilderModule {
    *
    * @return If the vertex has been connected.
    */
-  private boolean connectVertexToNode(TransitStopVertex ts, StreetIndex index) {
-    var nearbyBoardingLocations = index
-      .getVerticesForEnvelope(getEnvelope(ts))
+  private boolean connectVertexToNode(TransitStopVertex ts, Graph graph) {
+    var nearbyBoardingLocations = graph
+      .findVertices(getEnvelope(ts))
       .stream()
       .filter(OsmBoardingLocationVertex.class::isInstance)
       .map(OsmBoardingLocationVertex.class::cast)