+++ application/src/main/java/org/opentripplanner/updater/trip/siri/SiriRealTimeTripUpdateAdapter.java
@@ -29,6 +29,8 @@ import org.opentripplanner.updater.spi.UpdateError;
 import org.opentripplanner.updater.spi.UpdateResult;
 import org.opentripplanner.updater.spi.UpdateSuccess;
 import org.opentripplanner.updater.trip.TimetableSnapshotManager;
+import org.opentripplanner.updater.trip.TripPatternCache;
+import org.opentripplanner.updater.trip.TripPatternIdGenerator;
 import org.opentripplanner.updater.trip.UpdateIncrementality;
 import org.opentripplanner.utils.lang.StringUtils;
 import org.slf4j.Logger;
@@ -47,13 +49,12 @@ public class SiriRealTimeTripUpdateAdapter {
    * Use an id generator to generate TripPattern ids for new TripPatterns created by RealTime
    * updates.
    */
-  private final SiriTripPatternIdGenerator tripPatternIdGenerator =
-    new SiriTripPatternIdGenerator();
+  private final TripPatternIdGenerator tripPatternIdGenerator = new TripPatternIdGenerator();
   /**
    * A synchronized cache of trip patterns that are added to the graph due to GTFS-real-time
    * messages.
    */
-  private final SiriTripPatternCache tripPatternCache;
+  private final TripPatternCache tripPatternCache;
 
   /**
    * Long-lived transit editor service that has access to the timetable snapshot buffer.
@@ -66,17 +67,15 @@ public class SiriRealTimeTripUpdateAdapter {
 
   public SiriRealTimeTripUpdateAdapter(
     TimetableRepository timetableRepository,
-    TimetableSnapshotManager snapshotManager
+    TimetableSnapshotManager snapshotManager,
+    TripPatternCache tripPatternCache
   ) {
     this.snapshotManager = snapshotManager;
     this.transitEditorService = new DefaultTransitService(
       timetableRepository,
       snapshotManager.getTimetableSnapshotBuffer()
     );
-    this.tripPatternCache = new SiriTripPatternCache(
-      tripPatternIdGenerator,
-      transitEditorService::findPattern
-    );
+    this.tripPatternCache = tripPatternCache;
   }
 
   /**