+++ application/src/test/java/org/opentripplanner/routing/linking/LinkStopToPlatformTest.java
@@ -108,7 +108,7 @@ public class LinkStopToPlatformTest {
     }
 
     timetableRepository.index();
-    graph.index(timetableRepository.getSiteRepository());
+    graph.index();
 
     for (RegularStop s : transitStops) {
       vertexFactory.transitStop(TransitStopVertex.of().withStop(s));