+++ application/src/test/java/org/opentripplanner/service/paging/PS3_FewItinerariesOnSearchWindowLimitTest.java
@@ -8,6 +8,8 @@ import static org.opentripplanner.service.paging.TestPagingUtils.cleanStr;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -17,6 +19,7 @@ import org.opentripplanner.model.plan.ItinerarySortKey;
 import org.opentripplanner.model.plan.SortOrder;
 import org.opentripplanner.model.plan.paging.cursor.PageCursor;
 import org.opentripplanner.model.plan.paging.cursor.PageType;
+import org.opentripplanner.utils.time.DurationUtils;
 import org.opentripplanner.utils.time.TimeUtils;
 
 /**
@@ -49,33 +52,41 @@ class PS3_FewItinerariesOnSearchWindowLimitTest {
   private static final boolean ARRIVE_BY = true;
   public static final String EMPTY = "";
 
+  /**
+   * This matches strings like N5h1m and P1h. It is used to add a time shift value
+   * to the test cases that differs from the search window.
+   */
+  private static final Pattern PAGING_SEQUENCE_TIME_SHIFT_PATTERN = Pattern.compile("^(N|P)(.+)$");
+
   private final TestPagingModel model = TestPagingModel.testDataWithFewItinerariesCaseB();
   private TestDriver driver;
 
   /**
    * List of test-cases - for example:
    * <pre>
-   *   Given: "08:00", DEPART_AFTER, "1 N 2 N - P 2 -"
-   *     - "08:00"        : The first search departure time
-   *     - DEPART_AFTER   : If first search is arriveBy or depart after search
-   *     - "1 N 2 N - .." : Paging sequence - N:NEXT or P:PREVIOUS, 0-4 is the expected itinerary
-   *                        index found in the search between paging events, and '-' means no
-   *                        itinerary found.
+   *   Given: "08:00", DEPART_AFTER, "1 N3600 2 N - P 2 -"
+   *     - "08:00"            : The first search departure time
+   *     - DEPART_AFTER       : If first search is arriveBy or depart after search
+   *     - "1 N5h1m 2 N - .." : Paging sequence - N:NEXT or P:PREVIOUS, a duration directly after
+   *                            N or P indicates a time shift value that differs from the search
+   *                            window, individual integers in the range 0-4 are the expected
+   *                            itinerary index found in the search between paging events,
+   *                            and '-' means no itinerary found.
    * </pre>
    */
   static List<Arguments> testCases() {
     return List.of(
       // Itineraries depart inside search-window, step NEXT 6 times
-      //Arguments.of("08:00", DEPART_AFTER, "1 N 2 N - N - N 3 N - N"),
+      Arguments.of("08:00", DEPART_AFTER, "1 N 2 N - N - N 3 N - N"),
       // Itineraries depart inside search-window, step PREV 6 times
-      //Arguments.of("08:00+1d", DEPART_AFTER, "3 P - P - P 2 P 1 P - P"),
+      Arguments.of("08:00+1d", DEPART_AFTER, "3 P - P - P 2 P 1 P - P"),
       // Itineraries depart inside search-window, step BACK and FORTH
-      //Arguments.of("08:00", DEPART_AFTER, "1 N 2 N - P 2 N - P 2 P 1 N 2 N"),
+      Arguments.of("08:00", DEPART_AFTER, "1 N 2 N - P 2 N - P 2 P 1 N 2 N"),
       // Itineraries depart exactly at the start of search-window [inc, exc>, same sequence as
       // test above
       Arguments.of("09:00", DEPART_AFTER, "1 N 2 N - P 2 N - P 2 P 1 N 2 N"),
       // Itineraries depart inside search-window, step NEXT 6 times
-      Arguments.of("10:00-1d", ARRIVE_BY, "3 N - N - N 2 N 1 N - N"),
+      Arguments.of("10:00-1d", ARRIVE_BY, "3 N5h1m - N - N 2 N 1 N - N"),
       // Itineraries depart inside search-window, step PREV 6 times
       Arguments.of("08:00+1d", ARRIVE_BY, "0 P - P - P 1 P 2 P - P"),
       // Itineraries depart inside search-window, step BACK and FORTH
@@ -117,10 +128,10 @@ class PS3_FewItinerariesOnSearchWindowLimitTest {
       if (tc.gotoNewPage()) {
         if (tc.gotoPage() == NEXT_PAGE) {
           cursor = subject.nextPageCursor();
-          currTime += SEARCH_WINDOW_SEC;
+          currTime += tc.timeShift;
         } else {
           cursor = subject.previousPageCursor();
-          currTime -= SEARCH_WINDOW_SEC;
+          currTime -= tc.timeShift;
         }
         String description = tc.testDescription();
         assertEquals(tc.gotoPage(), cursor.type(), description);
@@ -149,27 +160,35 @@ class PS3_FewItinerariesOnSearchWindowLimitTest {
   }
 
   private TestCase parseToken(String token, StringBuilder sequence) {
-    switch (token) {
+    Matcher matcher = PAGING_SEQUENCE_TIME_SHIFT_PATTERN.matcher(token);
+    int timeShift = SEARCH_WINDOW_SEC;
+    String tokenWithTimeShiftRemoved = token;
+    if (matcher.matches()) {
+      tokenWithTimeShiftRemoved = matcher.group(1);
+      timeShift = (int) DurationUtils.duration(matcher.group(2)).toSeconds();
+    }
+    switch (tokenWithTimeShiftRemoved) {
       case "-":
         sequence.append(" > -");
-        return new TestCase(sequence.substring(3), null, -1, EMPTY);
+        return new TestCase(sequence.substring(3), null, -1, EMPTY, SEARCH_WINDOW_SEC);
       case "0", "1", "2", "3":
-        sequence.append(" > ").append(token);
-        int i = Integer.parseInt(token);
+        sequence.append(" > ").append(tokenWithTimeShiftRemoved);
+        int i = Integer.parseInt(tokenWithTimeShiftRemoved);
         return new TestCase(
           sequence.substring(3),
           null,
           i,
-          cleanStr(driver.all().get(i).keyAsString())
+          cleanStr(driver.all().get(i).keyAsString()),
+          SEARCH_WINDOW_SEC
         );
       case "N":
         sequence.append(" > NEXT");
-        return new TestCase(sequence.substring(3), NEXT_PAGE, -1, EMPTY);
+        return new TestCase(sequence.substring(3), NEXT_PAGE, -1, EMPTY, timeShift);
       case "P":
         sequence.append(" > PREV");
-        return new TestCase(sequence.substring(3), PREVIOUS_PAGE, -1, EMPTY);
+        return new TestCase(sequence.substring(3), PREVIOUS_PAGE, -1, EMPTY, timeShift);
       default:
-        throw new IllegalArgumentException(token);
+        throw new IllegalArgumentException(tokenWithTimeShiftRemoved);
     }
   }
 
@@ -181,7 +200,13 @@ class PS3_FewItinerariesOnSearchWindowLimitTest {
       .collect(Collectors.joining());
   }
 
-  record TestCase(String sequence, PageType gotoPage, int expIndex, String expItinerary) {
+  record TestCase(
+    String sequence,
+    PageType gotoPage,
+    int expIndex,
+    String expItinerary,
+    int timeShift
+  ) {
     boolean gotoNewPage() {
       return gotoPage != null;
     }